feat: add types and utilities for configuration management in Astro integrations
