+++ packages/@withstudiocms/config-utils/src/index.ts
@@ -1,3 +1,4 @@
+export * from './astro-integration-utils.js';
 export * from './loader.js';
 export * from './watcher.js';
 export * from './zod-utils.js';
