+++ packages/@withstudiocms/config-utils/src/loader.ts
@@ -1,11 +1,9 @@
 import { constants } from 'node:fs';
 import { access, unlink, writeFile } from 'node:fs/promises';
 import { fileURLToPath } from 'node:url';
-import type { z } from 'astro/zod';
-import { defineUtility } from 'astro-integration-kit';
 import { build as esbuild } from 'esbuild';
+import type { ImportBundledFileArgs, LoadAndBundleConfigFileArgs, LoadAndBundleConfigFileResult } from './types.js';
 import { tryCatch } from './utils/tryCatch.js';
-import { parseAndMerge, parseConfig } from './zod-utils.js';
 
 /**
  * Bundle arbitrary `mjs` or `ts` file.
@@ -39,12 +37,6 @@ export async function bundleConfigFile({ fileUrl }: { fileUrl: URL }) {
 	};
 }
 
-export interface ImportBundledFileArgs {
-	code: string;
-	root: URL;
-	label?: string; // Optional label for the temporary file
-}
-
 /**
  * Forked from Vite config loader, replacing CJS-based path concat with ESM only
  *
@@ -66,30 +58,6 @@ export async function importBundledFile({
 	} finally {
 		await tryCatch(unlink(tmpFileUrl));
 	}
-}
-
-/**
- * Arguments for loading and bundling a configuration file.
- *
- * @property root - The root directory as a URL.
- * @property fileUrl - The URL of the configuration file to load, or undefined if not specified.
- * @property label - (Optional) A label for the temporary file, used for identification or logging.
- */
-export interface LoadAndBundleConfigFileArgs {
-	root: URL;
-	fileUrl: URL | undefined;
-	label?: string; // Optional label for the temporary file
-}
-
-/**
- * The result of loading and bundling a configuration file.
- *
- * @property mod - The loaded module, which may contain a `default` export of any type, or be undefined if loading failed.
- * @property dependencies - An array of file paths representing the dependencies of the loaded configuration file.
- */
-export interface LoadAndBundleConfigFileResult {
-	mod: { default?: unknown } | undefined;
-	dependencies: string[];
 }
 
 /**
@@ -192,85 +160,3 @@ export async function loadConfigFile<R>(
 	// This is important for type safety and to ensure that the configuration object has the expected properties
 	return configMod.default as R;
 }
-
-/**
- * Options for building a configuration resolver.
- *
- * @template Schema - The type of the Zod schema used for validation.
- * @property configPaths - An array of file paths to search for configuration files.
- * @property label - A human-readable label describing the configuration.
- * @property zodSchema - The Zod schema instance used to validate the configuration.
- */
-export interface ConfigResolverBuilderOpts<Schema> {
-	configPaths: string[];
-	label: string;
-	zodSchema: Schema;
-}
-
-/**
- * Builds a config resolver utility for Astro integrations.
- *
- * This function creates a utility that loads, validates, and merges configuration
- * from both inline options and config files, using a provided Zod schema for validation.
- * If both inline config and a config file are present, the config file takes precedence during merging.
- *
- * @template ConfigType - The shape of the configuration object.
- * @template Schema - The Zod schema type used for validation.
- * @param params - The configuration resolver options.
- * @param params.configPaths - Array of possible config file paths to search for.
- * @param params.label - A label used for logging.
- * @param params.zodSchema - The Zod schema used to validate and parse the config.
- * @returns A utility function to be used in the Astro config setup hook, which loads, validates, and merges configuration.
- */
-export const configResolverBuilder = <S extends z.ZodTypeAny>({
-	configPaths,
-	label,
-	zodSchema,
-}: ConfigResolverBuilderOpts<S>) =>
-	defineUtility('astro:config:setup')(
-		async ({ logger: l, config: { root: astroRoot } }, options: S['_input']) => {
-			let inlineConfig: S['_output'] = {} as S['_output'];
-
-			// Generate a logger for the config resolver
-			const logger = l.fork(`${label}:config`);
-
-			// Check if inline options were provided
-			const inlineConfigExists = options !== undefined;
-
-			// If inline options are provided, parse them using the Zod schema
-			if (zodSchema) {
-				inlineConfig = parseConfig(zodSchema, options);
-			}
-
-			// Load the config file
-			const loadedConfigFile = await loadConfigFile<S['_input']>(astroRoot, configPaths, label);
-
-			// If no config file was found, return the inline config if it exists
-			if (!loadedConfigFile) {
-				logger.info('No config file found. Using inline config only.');
-				return inlineConfig;
-			}
-
-			// If inline config exists, log a warning
-			if (inlineConfigExists) {
-				logger.warn(
-					'Both an inline config and a config file were found. The config file will override the inline config during merging.'
-				);
-			}
-
-			// Parse and merge the inline config with the loaded config file
-			const mergedConfig = parseAndMerge(zodSchema, inlineConfig, loadedConfigFile);
-
-			let logMessage = 'Config file loaded and merged successfully.';
-
-			if (inlineConfigExists) {
-				logMessage += ` Warning: Inline config will be overridden by the config file, if you face any issues, try migrating your config to only use the ${label} config file.`;
-			}
-
-			logger.info(logMessage);
-
-			// Return the merged configuration object
-			// This object will contain the final configuration, validated and merged from both sources
-			return mergedConfig;
-		}
-	);
